♻️ (os-update): Moves os update commands to dmk ledger wallet
