+++ .changeset/smooth-trains-poke.md
@@ -1,0 +1,5 @@
+---
+"@ledgerhq/dmk-ledger-wallet": minor
+---
+
+Add OS update backup/restore commands and tasks (`BackupAppStorageCommand`, `GetAppStorageInfoCommand`, `CommitRestoreAppStorageCommand`, `InitRestoreAppStorageCommand`, `RequestMasterConsentCommand`, `RestoreAppStorageCommand`, `BackupAppStorageTask`, `RestoreAppStorageTask`), moved here from `@ledgerhq/device-management-kit`.
