+++ packages/ledger-wallet/src/index.ts
@@ -52,6 +52,45 @@ export {
 } from "./api/device-action/customLockScreenDeviceActionErrors";
 
 // Device Actions
+export {
+  BackupAppStorageCommand,
+  type BackupAppStorageCommandErrorCodes,
+  type BackupAppStorageCommandResponse,
+  type BackupAppStorageCommandResult,
+} from "./api/command/OsUpdate/Backup/BackupAppStorageCommand";
+export {
+  GetAppStorageInfoCommand,
+  type GetAppStorageInfoCommandArgs,
+  type GetAppStorageInfoCommandErrorCodes,
+  type GetAppStorageInfoCommandResponse,
+  type GetAppStorageInfoCommandResult,
+} from "./api/command/OsUpdate/Backup/GetAppStorageInfoCommand";
+export {
+  CommitRestoreAppStorageCommand,
+  CommitRestoreAppStorageCommandError,
+  type CommitRestoreAppStorageCommandErrorCodes,
+  type CommitRestoreAppStorageCommandResult,
+} from "./api/command/OsUpdate/Restore/CommitRestoreAppStorageCommand";
+export {
+  InitRestoreAppStorageCommand,
+  type InitRestoreAppStorageCommandArgs,
+  InitRestoreAppStorageCommandError,
+  type InitRestoreAppStorageCommandErrorCodes,
+  type InitRestoreAppStorageCommandResult,
+} from "./api/command/OsUpdate/Restore/InitRestoreAppStorageCommand";
+export {
+  RequestMasterConsentCommand,
+  type RequestMasterConsentCommandArgs,
+  type RequestMasterConsentCommandErrorCodes,
+  type RequestMasterConsentCommandResult,
+} from "./api/command/OsUpdate/Restore/RequestMasterConsentCommand";
+export {
+  RestoreAppStorageCommand,
+  type RestoreAppStorageCommandArgs,
+  RestoreAppStorageCommandError,
+  type RestoreAppStorageCommandErrorCodes,
+  type RestoreAppStorageCommandResult,
+} from "./api/command/OsUpdate/Restore/RestoreAppStorageCommand";
 export { DownloadCustomLockScreenDeviceAction } from "./api/device-action/DownloadCustomLockScreen/DownloadCustomLockScreenDeviceAction";
 export type {
   DownloadCustomLockScreenDAError,
@@ -98,3 +137,9 @@ export type {
   UploadCustomLockScreenDAState,
 } from "./api/device-action/UploadCustomLockScreen/types";
 export { UploadCustomLockScreenDeviceAction } from "./api/device-action/UploadCustomLockScreen/UploadCustomLockScreenDeviceAction";
+export { BackupAppStorageTask } from "./api/task/OsUpdate/Backup/BackupAppStorageTask";
+export {
+  RestoreAppStorageTask,
+  type RestoreAppStorageTaskArgs,
+  type RestoreAppStorageTaskError,
+} from "./api/task/OsUpdate/Restore/RestoreAppStorageTask";
